멜론 테스트 코드를 별도의 클래스로 분리
기존 MainTest 클래스에 있던 멜론 관련 테스트 코드를 별도의 GroupByTest 클래스로 분리하였습니다.

주요 변경사항:
1. MainTest 클래스에서 멜론 관련 테스트 코드를 제거하였습니다.
2. 새로운 GroupByTest 클래스를 생성하고, 여기에 멜론 관련 테스트 코드를 추가하였습니다.
